FEN checks also extract information
The FenString class not only checks the FEN string for validity, but
also extracts information like who is to move, the en passant target
square, and move numbers.
Not all information is extracted (e.g., the piece placement and castling
rights are missing). And the validity check for the en passant target
square is not yet dependent on the player to move next.